Harden diagnostic surfaces after quad-review
Addresses review feedback by warning on stale daemons that answer ping but reject status, sanitizing daemon-provided version strings before terminal output, using fixed-width JSON protocol counts, and covering negative log end ranges.

Constraint: Keep the fix narrow and avoid changing raw attach behavior or adding dependencies.
Rejected: Force-push amend | Local policy blocked force-push, so the PR keeps an explicit follow-up review-fix commit.
Confidence: high
Scope-risk: narrow
Directive: Treat daemon-provided status fields as untrusted for terminal output even over local IPC.
Tested: cargo fmt -- --check; cargo clippy --all-targets -- -D warnings; cargo test --test cli_smoke -- --test-threads=1; cargo test -- --test-threads=1; cargo build --release --locked; cargo audit; git diff --check
Not-tested: Live old-daemon binary compatibility beyond the Ping-succeeds/Status-fails warning path.

+++ src/protocol.rs
@@ -37,8 +37,8 @@ pub struct SessionInfo {
 pub struct DaemonStatus {
     pub version: String,
     pub protocol_version: u32,
-    pub session_count: usize,
-    pub active_connections: usize,
+    pub session_count: u64,
+    pub active_connections: u64,
     pub shutting_down: bool,
 }
